Refactor price fields to use BigDecimal for monetary values across DTOs and entities; update validation and service logic accordingly. Add Postman collection for cart and checkout flow testing.

+++ backend_Order/src/main/java/com/example/backend_Ecom/dto/CartResponseDto.java
@@ -2,6 +2,7 @@ package com.example.backend_Ecom.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -16,7 +17,7 @@ public class CartResponseDto {
 
     private Long userId;
 
-    private Long totalPrice;
+    private BigDecimal totalPrice;
 
     private List<CartItemResponseDto> items;
